backend/pyproject.toml
[project]
name = "itpe-enhancement-backend"
version = "1.0.0"
description = "ITPE Topic Enhancement System Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "alembic>=1.13.0",
    "pdfplumber>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "numpy>=2.1.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.9",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
# SQLite 동시성 문제로 인해 테스트 순차 실행
addopts = "-v --tb=short -p no:xdist"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
